+++ src/graphics/texture.rs
@@ -26,9 +26,14 @@ impl PartialEq for TextureSharedData {
 
 /// A texture, held in GPU memory.
 ///
-/// # Supported Formats
+/// The data can be stored in a variety of formats, as represented by the
+/// [`TextureFormat`] enum.
 ///
-/// Various file formats are supported, and can be enabled or disabled via Cargo features:
+/// # Supported File Formats
+///
+/// Images can be decoded from various common file formats via the [`new`](Texture::new)
+/// and [`from_encoded`](Texture::from_encoded) constructors. Individual
+/// decoders can be enabled or disabled via Cargo feature flags.
 ///
 /// | Format | Cargo feature | Enabled by default? |
 /// |-|-|-|
@@ -68,9 +73,12 @@ impl Texture {
     ///
     /// # Errors
     ///
-    /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
-    /// * [`TetraError::FailedToLoadAsset`] will be returned if the file could not be loaded.
-    /// * [`TetraError::InvalidTexture`] will be returned if the texture data was invalid.
+    /// * [`TetraError::PlatformError`](crate::TetraError::PlatformError) will be returned
+    /// if the underlying graphics API encounters an error.
+    /// * [`TetraError::FailedToLoadAsset`](crate::TetraError::FailedToLoadAsset) will be
+    /// returned if the file could not be loaded.
+    /// * [`TetraError::InvalidTexture`](crate::TetraError::InvalidTexture) will be returned
+    /// if the texture data was invalid.
     pub fn new<P>(ctx: &mut Context, path: P) -> Result<Texture>
     where
         P: AsRef<Path>,
@@ -89,9 +97,11 @@ impl Texture {
     ///
     /// # Errors
     ///
-    /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
-    /// * [`TetraError::NotEnoughData`] will be returned if not enough data is provided to fill
-    /// the texture. This is to prevent the graphics API from trying to read uninitialized memory.
+    /// * [`TetraError::PlatformError`](crate::TetraError::PlatformError) will be returned
+    /// if the underlying graphics API encounters an error.
+    /// * [`TetraError::NotEnoughData`](crate::TetraError::NotEnoughData) will be returned
+    /// if not enough data is provided to fill the texture. This is to prevent the
+    /// graphics API from trying to read uninitialized memory.
     pub fn from_data(
         ctx: &mut Context,
         width: i32,
@@ -122,8 +132,10 @@ impl Texture {
     ///
     /// # Errors
     ///
-    /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
-    /// * [`TetraError::InvalidTexture`] will be returned if the texture data was invalid.
+    /// * [`TetraError::PlatformError`](crate::TetraError::PlatformError) will be
+    /// returned if the underlying graphics API encounters an error.
+    /// * [`TetraError::InvalidTexture`](crate::TetraError::InvalidTexture) will be
+    /// returned if the texture data was invalid.
     pub fn from_encoded(ctx: &mut Context, data: &[u8]) -> Result<Texture> {
         let data = ImageData::from_encoded(data)?;
         Texture::from_image_data(ctx, &data)
@@ -133,7 +145,8 @@ impl Texture {
     ///
     /// # Errors
     ///
-    /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
+    /// * [`TetraError::PlatformError`](crate::TetraError::PlatformError) will be returned
+    /// if the underlying graphics API encounters an error.
     pub fn from_image_data(ctx: &mut Context, data: &ImageData) -> Result<Texture> {
         Texture::from_data(
             ctx,
@@ -344,10 +357,13 @@ impl Texture {
     /// or if you want to output the image data somewhere. This is a fairly
     /// slow operation, so avoid doing it too often!
     pub fn get_data(&self, ctx: &mut Context) -> ImageData {
+        // TODO: Should this allow copying to a different format?
+
         let (width, height) = self.size();
         let buffer = ctx.device.get_texture_data(&self.data.handle);
 
-        ImageData::from_data(width, height, buffer).expect("buffer should be exact size for image")
+        ImageData::from_data(width, height, TextureFormat::Rgba8, buffer)
+            .expect("buffer should be exact size for image")
     }
 
     /// Writes RGBA pixel data to a specified region of the texture.
@@ -361,9 +377,9 @@ impl Texture {
     ///
     /// # Errors
     ///
-    /// * [`TetraError::NotEnoughData`] will be returned if not enough data is provided to fill
-    /// the target rectangle. This is to prevent the graphics API from trying to read
-    /// uninitialized memory.
+    /// * [`TetraError::NotEnoughData`](crate::TetraError::NotEnoughData) will be returned if
+    /// not enough data is provided to fill the target rectangle. This is to prevent the
+    /// graphics API from trying to read uninitialized memory.
     ///
     /// # Panics
     ///
@@ -392,8 +408,9 @@ impl Texture {
     ///
     /// # Errors
     ///
-    /// * [`TetraError::NotEnoughData`] will be returned if not enough data is provided to fill
-    /// the texture. This is to prevent the graphics API from trying to read uninitialized memory.
+    /// * [`TetraError::NotEnoughData`](crate::TetraError::NotEnoughData) will be returned if not
+    /// enough data is provided to fill the texture. This is to prevent the graphics API from
+    /// trying to read uninitialized memory.
     pub fn replace_data(&self, ctx: &mut Context, data: &[u8]) -> Result {
         let (width, height) = self.size();
         self.set_data(ctx, 0, 0, width, height, data)
@@ -416,6 +433,17 @@ pub enum TextureFormat {
 impl TextureFormat {
     /// Returns the number of channels that the format has.
     pub fn channels(self) -> usize {
+        match self {
+            TextureFormat::Rgba8 => 4,
+            TextureFormat::R8 => 1,
+            TextureFormat::Rg8 => 2,
+        }
+    }
+
+    /// Returns the number of bytes per pixel for this format.
+    ///
+    /// TODO: Should this be public?
+    pub(crate) fn stride(self) -> usize {
         match self {
             TextureFormat::Rgba8 => 4,
             TextureFormat::R8 => 1,